application/backend/pyproject.toml
[project]
name = "anomalib-studio"
version = "1.0.0"
description = "Anomalib Studio server"
requires-python = ">=3.13"

dependencies = [
    "aiofiles==25.1.0",
    "aiortc>=1.13.0",
    "aiosqlite~=0.21",
    "alembic~=1.18.1",
    "cv2-enumerate-cameras==1.3.3",
    "fastapi[standard]~=0.116",
    "jsonschema>=4.18.0a10",
    "loguru==0.7.3",
    "openvino==2025.4.1",
    "opencv-python-headless~=4.12",
    "psutil~=7.2.1",
    "pyarrow~=23.0",
    "pydantic-settings~=2.12.0",
    "requests",
    "shortuuid~=1.0",
    "sqlalchemy[asyncio]~=2.0",
    "sse-starlette~=3.0",
    "tensorboard",
    "uvicorn~=0.17",
    "uvloop==0.22.1 ; sys_platform != 'win32'",
    "watchdog==6.0.0",
]

[dependency-groups]
dev = [
    "pytest~=9.0",
    "prek",
    "pdbpp~=0.10",
    "flaky~=3.8",
    "testfixtures~=10.0",
    "httpx~=0.24",
    "freezegun~=1.5",
    "pytest-asyncio~=1.3.0",
]
lint = [
    "ruff~=0.11.2",
    "mypy==1.19.1",
    "types-requests~=2.32.4",
    "types-PyYAML~=6.0.12",
]

[project.optional-dependencies]
mqtt = [
    "paho-mqtt~=2.1.0",
]

# Explicit versions are needed as we can't propagate extra dependencies to anomalib
cpu = ["anomalib[cpu,openvino]"]
cu130 = ["anomalib[cu130,openvino]"]
xpu = ["anomalib[xpu,openvino] ; sys_platform == 'linux' or sys_platform == 'win32'"]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cu130" },
        { extra = "xpu" },
    ],
    # redeclare library conflicts to workaround bug with uv dependency resolution
    # For details, see: https://github.com/astral-sh/uv/issues/9942
    [
        { package = "anomalib", extra = "cpu" },
        { package = "anomalib", extra = "cu130" },
        { package = "anomalib", extra = "xpu" },
    ],
]

[tool.uv.sources]
anomalib = { path = "../../", editable = true }


[tool.ruff]
extend = "../../pyproject.toml"
target-version = "py313"
line-length = 120

exclude = [".venv*"]

src = ["src"]

lint.select = [
    # --- Core & Standard Rules ---
    "E",      # pycodestyle (Error)
    "F",      # Pyflakes (General logic/syntax errors)
    "I",      # isort (Import sorting)
    "UP",     # pyupgrade (Upgrades syntax to newer Python versions)
    "N",      # pep8-naming (Naming conventions)

    # --- Type Hinting & Annotations ---
    "ANN001", # Missing type annotation for function argument
    "ANN201", # Missing return type annotation for public function
    "ANN205", # Missing return type annotation for static method
    "TC",     # flake8-type-checking (Type-check blocks)
    "PYI",    # flake8-pyi (Stub file linting)
    "FA",     # flake8-future-annotations (Future import for types)

    # --- Security & Async ---
    "S",      # flake8-bandit (Security testing)
    "ASYNC",  # flake8-async (Async-specific linting)

    # --- Best Practices & Simplification ---
    "ARG",    # flake8-unused-arguments (Unused function arguments)
    "C4",     # flake8-comprehensions (List/set/dict comprehension fixes)
    "SIM",    # flake8-simplify (Simplify complex code)
    "RET",    # flake8-return (Return statement consistency)
    "PIE",    # flake8-pie (Miscellaneous linting)
    "RUF",    # Ruff-specific rules
    "PL",     # Pylint (Refactor, Error, Warning, Convention)

    # --- Style & Formatting ---
    "COM",    # flake8-commas (Trailing commas)
    "Q",      # flake8-quotes (Consistent string quotes)
    "C90",    # mccabe (Code complexity/Cyclomatic complexity)
    "D103",   # pydocstyle (Missing docstring in public function)
    "RSE",    # flake8-raise (Specifics on raising exceptions)
    "TID",    # flake8-tidy-imports (Clean up messy imports)

    # --- Ecosystem Specific ---
    "FAST",   # FastAPI (Rules specific to FastAPI apps)
    "YTT",    # flake8-2020 (Checks for system versioning issues)
]
lint.ignore = [
    # --- Naming Conventions (pep8-naming) ---
    "N801",   # Class name should use CapWords convention
    "N805",   # First argument of a method should be named 'self'
    "N806",   # Variable in a function should be lowercase
    "N807",   # Function name should not start and end with '__'
    "N818",   # Exception name should be named with an 'Error' suffix

    # --- Code Simplification & Style ---
    "SIM105", # Use contextlib.suppress instead of try-except-pass
    "SIM108", # Use ternary operator instead of if-else block
    "RET503", # Missing explicit return at the end of function
    "COM812", # Missing trailing comma in a multi-line collection

    # --- Pylint Refactor (PLR) ---
    "PLR2004", # Magic value used in comparison (e.g., status == 200)
    "PLR6301", # Method could be a function/static (doesn't use 'self')

    # --- Type Checking & Ruff Specifics ---
    "RUF010", # Use explicit conversion flags (!r, !s) in f-strings
    "RUF012", # Mutable class attributes should be annotated with ClassVar
    "RUF029", # Unused 'async' keyword on a function
]
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.pytest.ini_options]
pythonpath = ["src"]
filterwarnings = [
    "ignore:.*fork.*may lead to deadlocks.*:DeprecationWarning:multiprocessing",
    "ignore:Please use.*python_multipart.*:PendingDeprecationWarning:starlette",
    "ignore:Importing from timm.models.layers is deprecated.*:FutureWarning:timm",
    # Suppress AsyncMock coroutine warnings - these are test artifacts from unittest.mock
    # framework internals during garbage collection, not actual code issues
    "ignore:coroutine.*was never awaited:RuntimeWarning:unittest.mock",
    # Suppress websockets deprecation warnings from uvicorn (third-party dependency issue)
    "ignore:.*websockets.legacy is deprecated.*:DeprecationWarning",
    "ignore:.*WebSocketServerProtocol is deprecated.*:DeprecationWarning",
    # Suppress third-party library deprecation warnings
    "ignore:Jupyter is migrating its paths.*:DeprecationWarning:jupyter_client",
    "ignore:.*sentry_sdk.Hub.*is deprecated.*:DeprecationWarning",
]

[tool.ruff.lint.per-file-ignores]
"*test*.py" = ["S", "ANN", "D", "ARG", "PLR"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pylint]
max-args=7

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
show_error_codes = true
